+++ src/targets/gpu/include/migraphx/gpu/compile_pointwise.hpp
@@ -16,9 +16,9 @@ compile_pointwise(context& ctx,
                   const std::vector<shape>& inputs,
                   const std::string& lambda,
                   const std::string& preamble = "",
-                  int global_workitems = 1024, // Global (total) work items. todo: testing indicates
+                  size_t global_workitems = 1024, // Global (total) work items. todo: testing indicates
                                                // values of global close to tensor size are best
-                  int local_workitems = 256);  // Local work items per CU
+                  size_t local_workitems = 256);  // Local work items per CU
 operation compile_pointwise(context& ctx, const std::vector<shape>& inputs, module m);
 
 } // namespace gpu
